+++ lib/src/bones_api_base.dart
@@ -41,7 +41,7 @@ typedef APILogger = void Function(APIRoot apiRoot, String type, String? message,
 /// Bones API Library class.
 class BonesAPI {
   // ignore: constant_identifier_names
-  static const String VERSION = '1.4.28';
+  static const String VERSION = '1.4.29';
 
   static bool _boot = false;